Extract layer activate to drawSelect
